+++ methods/person.rb
