+++ methods/student.rb
